perf

+++ src/gameEngine/inputEngine/input.cpp
@@ -62,7 +62,7 @@ static inline void updateInput()
                 SDL_GetWindowSize(window, &windowwidth, &windowheight);
                 glViewport(0, 0, windowwidth, windowheight);
 
-                const float ratio = windowwidth / windowheight;
+                const float ratio = float((float)windowwidth / (float)windowheight);
                 const float scale = 700.0;
                 mainCamera.orthoProjectionMatrix = glm::ortho(-windowwidth / scale, windowwidth / scale, -windowheight / scale, windowheight / scale, -1.0f, 100.0f);
